Change project structure.
