--- Client/core/src/com/mygdx/client/model/Bomb.java
@@ -1,4 +1,0 @@
-package com.mygdx.client.model;
-
-public class Bomb extends MapObject {
-}
