--- Client/core/src/com/mygdx/client/model/MapObject.java
@@ -1,8 +1,0 @@
-package com.mygdx.client.model;
-
-public class MapObject {
-    private int posX;
-    private int posY;
-    int getPosX() { return posX; }
-    int getPosY() { return posY; }
-}
