--- Client/core/src/com/mygdx/client/model/Player.java
@@ -1,4 +1,0 @@
-package com.mygdx.client.model;
-
-public class Player extends MapObject {
-}
