--- Client/core/src/com/mygdx/client/model/PowerUp.java
@@ -1,4 +1,0 @@
-package com.mygdx.client.model;
-
-public class PowerUp extends MapObject {
-}
