--- Client/core/src/com/mygdx/client/model/Wall.java
@@ -1,5 +1,0 @@
-package com.mygdx.client.model;
-
-public class Wall extends MapObject {
-
-}
