--- Client/core/src/com/mygdx/client/serialization/Parser.java
@@ -1,16 +1,0 @@
-package com.mygdx.client.serialization;
-
-import com.mygdx.client.model.Map;
-
-/**
- * TODO
- * Parser pobiera z pliku w folderze "assets" plik tekstowy
- * mapy i zamienia znaki na obiekty
- */
-public class Parser {
-
-    public Map readMapFromFile() {
-
-        return null;
-    }
-}
