+++ Client/src/com/client/game/DesktopLauncher.java
@@ -1,0 +1,15 @@
+package com.client.game;
+
+import com.badlogic.gdx.backends.lwjgl3.Lwjgl3Application;
+import com.badlogic.gdx.backends.lwjgl3.Lwjgl3ApplicationConfiguration;
+import com.core.game.Bomberman;
+
+// Please note that on macOS your application needs to be started with the -XstartOnFirstThread JVM argument
+public class DesktopLauncher {
+	public static void main (String[] arg) {
+		Lwjgl3ApplicationConfiguration config = new Lwjgl3ApplicationConfiguration();
+		config.setForegroundFPS(60);
+		config.setTitle("Bomberman");
+		new Lwjgl3Application(new Bomberman(), config);
+	}
+}
