--- Server/gradle/wrapper/gradle-wrapper.properties
@@ -1,6 +1,0 @@
-#Sun May 21 21:45:52 CEST 2023
-distributionBase=GRADLE_USER_HOME
-distributionPath=wrapper/dists
-distributionUrl=https\://services.gradle.org/distributions/gradle-7.5.1-bin.zip
-zipStoreBase=GRADLE_USER_HOME
-zipStorePath=wrapper/dists
