+++ Server/src/com/server/game/Server.java
@@ -1,0 +1,14 @@
+package com.server.game;
+
+import com.core.game.model.Map;
+
+public class Server {
+    public static void main(String[] args) {
+
+        System.out.println("Hello world!");
+        Map map = new Map();
+        //map.loadMapFromFile("map.txt");
+
+        int p;
+    }
+}
