--- Server/src/main/java/Shared/Bomb.java
@@ -1,9 +1,0 @@
-package Shared;
-
-public class Bomb extends MapObject {
-
-    public Bomb(int positionX, int positionY) {
-        super(positionX, positionY, false, false);
-    }
-
-}
