--- Server/src/main/java/Shared/Floor.java
@@ -1,9 +1,0 @@
-package Shared;
-
-public class Floor extends MapObject {
-
-    public Floor(int positionX, int positionY) {
-        super(positionX, positionY, false, true);
-    }
-
-}
