--- Server/src/main/java/Shared/Player.java
@@ -1,12 +1,0 @@
-package Shared;
-
-public class Player extends MapObject {
-
-    private int playerID;
-
-    public Player(int positionX, int positionY, int playerID) {
-        super(positionX, positionY, true, true);
-        this.playerID = playerID;
-    }
-
-}
