--- Server/src/main/java/Shared/PowerUP.java
@@ -1,9 +1,0 @@
-package Shared;
-
-public class PowerUP extends MapObject{
-
-    public PowerUP(int positionX, int positionY){
-        super(positionX,positionY,false,true);
-    }
-
-}
