--- Server/src/main/java/Shared/Wall.java
@@ -1,9 +1,0 @@
-package Shared;
-
-public class Wall extends MapObject {
-
-    public Wall(int positionX, int positionY, boolean destructible) {
-        super(positionX, positionY, destructible, false);
-    }
-
-}
