--- Server/src/main/java/org/example/Server.java
@@ -1,7 +1,0 @@
-package org.example;
-
-public class Server {
-    public static void main(String[] args) {
-        System.out.println("Hello world!");
-    }
-}
