+++ assets/map.txt
@@ -1,0 +1,8 @@
+-------------
+-------------
+-----xxxx----
+-----+++-----
+-------------
+-------------
+-------------
+-------------
