+++ core/src/com/core/game/model/Bomb.java
@@ -1,0 +1,9 @@
+package com.core.game.model;
+
+public class Bomb extends MapObject {
+
+    public Bomb(int positionX, int positionY) {
+        super(positionX, positionY, false, false);
+    }
+
+}
