+++ core/src/com/core/game/model/Floor.java
@@ -1,0 +1,7 @@
+package com.core.game.model;
+
+public class Floor extends MapObject {
+    public Floor(int positionX, int positionY) {
+        super(positionX, positionY, false, true);
+    }
+}
