+++ core/src/com/core/game/model/Player.java
@@ -1,0 +1,12 @@
+package com.core.game.model;
+
+public class Player extends MapObject {
+
+    private int playerID;
+
+    public Player(int positionX, int positionY, int playerID) {
+        super(positionX, positionY, true, true);
+        this.playerID = playerID;
+    }
+
+}
