+++ core/src/com/core/game/model/PowerUp.java
@@ -1,0 +1,10 @@
+package com.core.game.model;
+
+
+public class PowerUp extends MapObject{
+
+    public PowerUp(int positionX, int positionY){
+        super(positionX,positionY,false,true);
+    }
+
+}
