+++ core/src/com/core/game/model/Wall.java
@@ -1,0 +1,9 @@
+package com.core.game.model;
+
+public class Wall extends MapObject {
+
+    public Wall(int positionX, int positionY, boolean destructible) {
+        super(positionX, positionY, destructible, false);
+    }
+
+}
