+++ core/src/com/core/game/serialization/Parser.java
@@ -1,0 +1,16 @@
+package com.core.game.serialization;
+
+import com.core.game.model.Map;
+
+/**
+ * TODO
+ * Parser pobiera z pliku w folderze "assets" plik tekstowy
+ * mapy i zamienia znaki na obiekty
+ */
+public class Parser {
+
+    public Map readMapFromFile() {
+
+        return null;
+    }
+}
